[FIX] Adding DropdownButton and UploadImageButton, removing SettingsButton
