+++ skill-db/frontend/src/components/DetailComponent.vue
@@ -1,4 +1,10 @@
 <template>
+  <div class="d-flex justify-end">
+    <dropdown-button :functions="dropdownFunctions"/>
+    <delete-profile-dialog v-model="toDelete" :functions="dialogFunctions"/>
+    <v-overlay v-model="locked" absolute/>
+  </div>
+
   <div class="header">
     <img src="@/assets/images/dummy_profilePicture.png" alt="Profilbild">
     <div class="header_content d-flex flex-column align-content-space-between">
@@ -79,9 +85,56 @@
 </template>
 
 <script lang="ts">
+import DropdownButton from "@/components/DropdownButton.vue";
+import DeleteProfileDialog from "@/components/DeleteProfileDialog.vue";
+import {useDetailStore} from "@/stores/DetailStore";
+import {useRoute} from "vue-router";
+import {ref} from "vue";
+import router from "@/router";
+
 export default {
   name: "DetailComponent",
-  props: ['detailStore'],
+  components: { DropdownButton, DeleteProfileDialog },
+  setup() {
+    const detailStore = useDetailStore();
+    const id = Number(useRoute().params.id);
+    detailStore.loadDetailsById(id);
+
+    const locked = ref(false);
+    const toDelete = ref(false);
+
+    function enterEdit(): void {
+      router.push({name: 'editView', params: {id: id}});
+    }
+
+    function toggleDelete(): void {
+      toDelete.value = !toDelete.value;
+    }
+
+    function lockProfile(): void {
+      locked.value = true;
+      console.log("This profile is now locked away.");
+      /*router.push(`/`);*/
+    }
+
+    function deleteProfile(): void {
+      detailStore.deleteDetailsByID(id);
+      router.push(`/`);
+    }
+
+    return {
+      dropdownFunctions: [
+        {name: 'Bearbeiten', method: enterEdit},
+        {name: 'Löschen', method: toggleDelete},
+        {name: 'Sperren', method: lockProfile},
+      ],
+      dialogFunctions: [
+        {name: 'Ja', method: deleteProfile},
+        {name: 'Nein', method: toggleDelete}
+      ],
+      detailStore, toDelete, locked
+    };
+  }
 }
 </script>
