+++ skill-db/frontend/src/views/EditView.vue
@@ -7,8 +7,6 @@
 
 <script lang="ts">
 import EditComponent from "@/components/EditComponent.vue";
-import {ref} from "vue";
-import router from "@/router";
 import {useDetailStore} from "@/stores/DetailStore";
 import {useRoute} from "vue-router";
 
@@ -20,27 +18,8 @@ export default {
     const id = Number(useRoute().params.id);
     detailStore.loadDetailsById(id);
 
-    const locked = ref(false);
-    const toDelete = ref(false);
-
-    function toggleDelete(): void {
-      toDelete.value = !toDelete.value;
-    }
-
-    function lockProfile(): void {
-      locked.value = true;
-      console.log("This profile is now locked away.");
-      /*router.push(`/`);*/
-    }
-
-    function deleteProfile(): void {
-      detailStore.deleteDetailsByID(id);
-      router.push(`/`);
-    }
-
     return {
-      detailStore, toDelete, locked, toggleDelete,
-      lockProfile, deleteProfile
+      detailStore
     };
   },
 }
